#272 Rename neighborhood bbox to canonical bbox entrypoints
Per Table 4.2.1 the neighborhood-scale bbox row is THE bbox row, so the
neighborhood entrypoints become the canonical bbox entrypoints:
bbox_filtering_result_set_sizes_neighborhood_{duckdb,postgis,local}.py
-> bbox_filtering_{duckdb,postgis,local}.py. All three switch to the
factory pattern. The local variant raises ValueError when dataset_size
is not SMALL, since the Shapefile target is small-only per Table 4.2.1.
The __init__ re-exports, app_config wiring, benchmark_runner dispatch,
and benchmarks.yml entries are updated to match the new names.
