+++ src/presentation/entrypoints/__init__.py
@@ -6,13 +6,13 @@ from .bbox_filtering_advanced_duckdb import bbox_filtering_advanced_duckdb
 from .bbox_filtering_advanced_postgis import bbox_filtering_advanced_postgis
 from .bbox_filtering_simple_local import bbox_filtering_simple_local
 from .bbox_filtering_simple_blob_storage import bbox_filtering_simple_blob_storage
-from .bbox_filtering_result_set_sizes_neighborhood_duckdb import bbox_filtering_result_set_sizes_neighborhood_duckdb
+from .bbox_filtering_duckdb import bbox_filtering_duckdb
 from .bbox_filtering_result_set_sizes_municipality_duckdb import bbox_filtering_result_set_sizes_municipality_duckdb
 from .bbox_filtering_result_set_sizes_county_duckdb import bbox_filtering_result_set_sizes_county_duckdb
-from .bbox_filtering_result_set_sizes_neighborhood_postgis import bbox_filtering_result_set_sizes_neighborhood_postgis
+from .bbox_filtering_postgis import bbox_filtering_postgis
 from .bbox_filtering_result_set_sizes_municipality_postgis import bbox_filtering_result_set_sizes_municipality_postgis
 from .bbox_filtering_result_set_sizes_county_postgis import bbox_filtering_result_set_sizes_county_postgis
-from .bbox_filtering_result_set_sizes_neighborhood_local import bbox_filtering_result_set_sizes_neighborhood_local
+from .bbox_filtering_local import bbox_filtering_local
 from .bbox_filtering_result_set_sizes_municipality_local import bbox_filtering_result_set_sizes_municipality_local
 from .bbox_filtering_result_set_sizes_county_local import bbox_filtering_result_set_sizes_county_local
 from .ordered_range_query_duckdb import ordered_range_query_duckdb
